pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "consult7"
version = "3.0.0"
description = "MCP server for consulting large context window models to analyze extensive file collections via OpenRouter"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Stefan Szeider", email = "stefan@szeider.net"}
]
keywords = ["mcp", "llm", "code-analysis", "large-context", "openrouter"]
dependencies = [
    "mcp>=1.9.4",
    "httpx>=0.28.1",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "openai>=1.59.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.1",
]

[project.urls]
Homepage = "https://github.com/szeider/consult7"
Repository = "https://github.com/szeider/consult7"
Issues = "https://github.com/szeider/consult7/issues"

[project.scripts]
consult7 = "consult7.server:run"

[tool.hatch.build.targets.wheel]
packages = ["src/consult7"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "anthropic>=0.52.0",  # For MCP testing support
    "pytest>=8.3.5",  # Testing framework
    "pytest-asyncio>=1.3.0",  # Async test support
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "T201",   # flake8-print: print found
]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.lint.per-file-ignores]
# Allow print statements in test files
"tests/**/*.py" = ["T201"]
# Allow print statements in server.py for CLI output and test mode
"src/consult7/server.py" = ["T201"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_default_fixture_loop_scope = "function"
